Fix the narrowing conversion that broke the per-pixel build
agiRendStateStruct::MaxTextures is an i8 (agi/rsys.h:260), so handing
SetMaxTextures an i32 is C4244, which is an error under /W4 /WX. The value
is 1 or 2 and cannot overflow; the cast just says so.

+++ code/midtown/agidx9/dx9pipe.cpp
@@ -156,7 +156,9 @@ i32 agiDX9Pipeline::BeginGfx()
     // nothing could use more.
     if (per_pixel_.IsValid())
     {
-        agiCurState.SetMaxTextures(std::min<i32>(static_cast<i32>(dx9_context_->GetMaxSimultaneousTextures()), 2));
+        // agiRendStateStruct::MaxTextures is an i8, so the narrowing is explicit - the value is 1 or
+        // 2 and cannot overflow, but /W4 /WX rejects the implicit conversion.
+        agiCurState.SetMaxTextures(static_cast<i8>(std::min<u32>(dx9_context_->GetMaxSimultaneousTextures(), 2u)));
     }
 
     InitScaling();
